+++ changelog.md
@@ -2,8 +2,10 @@
 
 ## **2026-XX-XX** - 0.0.7
 
-1. REMOVED `extract` mod.
-1. MODIFIED `osm_pbf_file` domain.
+1. REMOVED `extract/osm_data` mod.
+1. ADDED `osm_node`, `osm_way` and `osm_relation` domains with their decoders.
+1. MODIFIED `osm_pbf_file` domain, adding `extract_blob_chunks`, `extract_osm_header` and `extract_osm_data` to its facade.
+1. MODIFIED `database` mod, adding the `jsonb` encoder moved from `extract/osm_data`.
 
 ## **2026-09-06** - 0.0.6